+++ gdtoolkit/formatter/expression_to_str.py
@@ -102,6 +102,10 @@ def expression_to_str(expression: Node) -> str:
         "enum_body": _enum_body_to_str,
         "enum_element": _enum_element_to_str,
         "signal_args": _args_to_str,
+        "signal_arg_typed": lambda e: "{}: {}".format(
+            e.children[0].value,
+            e.children[1].value,
+        ),
         "comma_separated_list": lambda e: _arguments_to_str(e.children),
         "contextless_comma_separated_list": lambda e: _arguments_to_str(e.children),
         "contextless_operator_chain_based_expression": (