+++ tests/formatter/input-output-pairs/complex-signal-statements.in.gd
@@ -2,3 +2,7 @@ class X:
 	signal a(x, y)
 	signal b(aaaaaaaa, bbbbbbbbb, ccccccccc, dddddddd, eeeeeeee, ffffffff, ggggggggg, hhhhhhhh, iiiiiiii, jjjjjjjjjjj)
 	signal c(x, y,)
+	signal d(x:int,)
+	signal e(x,y:float)
+	signal f(x:int,aaaaaaaa, bbbbbbbbb, ccccccccc, dddddddd, eeeeeeee, ffffffff, ggggggggg, hhhhhhhh, iiiiiiii, jjjjjjjjjjj,y:float)
+	signal g(xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx: int)