+++ tests/formatter/input-output-pairs/type-test-expressions.in.gd
@@ -2,3 +2,4 @@ class X:
 	func foo():
 		var x = 1 is int is bool
 		var y = 1 is int is bool is bool is bool is bool is bool is bool is bool is bool is bool is bool is bool is bool is bool
+		print([1,2.1] is Array[int])