+++ tests/formatter/input-output-pairs/bug-205-static-func-corner-case.out.gd
@@ -1,0 +1,10 @@
+static func _is_agent_placement_position_valid(
+	position, radius, existing_units, navigation_map_rid
+):
+	pass
+
+
+func _is_agent_placement_position_valid2(
+	position, radius, existing_units, navigation_map_riddddddd
+):
+	pass