+++ tests/formatter/input-output-pairs/type-cast-corner-case-expressions.in.gd
@@ -34,3 +34,6 @@ func foo():
 	# ---
 	1 as int or 1 is int1
 	1 as int or [1,] is Array
+	# --
+	1 as int or 1**1
+	1 as int or [1,]**2