+++ tests/valid-gd-scripts/expressions.gd
@@ -42,6 +42,7 @@ func foo():
 	6 % 3
 	-8
 	~8
+	5 ** 5
 	x is int
 	x is Xyz # TODO: fix/remove x.Type
 	x is Zyx.Qwe # TODO: fix/remove x.Type