+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
  - Added support for user-defined types in type hints
  - Added support for annotations with empty argument list (e.g. `@onready()`)
+ - Added support for power operator `**`
 
 ### Fixed
  - Fixed `max-returns` check's message